new atoms module for parsing atoms
